# docs/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Performance Observability Hub API
  version: 0.1.0
  summary: API contract for multi-site performance ingestion and dashboard reads
servers:
  - url: https://perf.example.com
...
paths:
...
  /api/v1/dashboard/overview:
    get:
      tags: [Dashboard]
      summary: Return portfolio-level health and trend slices
      operationId: getOverview
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: environment
          schema:
            $ref: "#/components/schemas/Environment"
      responses:
        "200":
          description: Overview payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewResponse"
...
  /api/v1/sites/{siteId}/deployments/compare:
    get:
      tags: [Dashboard]
      summary: Compare two deployments or one deployment versus previous
      operationId: compareDeployments
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: currentDeploymentId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: baselineDeploymentId
          schema:
            type: string
            format: uuid
        - in: query
          name: deviceClass
          schema:
            $ref: "#/components/schemas/DeviceClass"
      responses:
        "200":
          description: Deployment comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentCompareResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SiteIngestKey:
      type: apiKey
      in: header
      name: X-Site-Ingest-Key
...
  schemas:
    Environment:
      type: string
      enum: [production, staging, preview, development]
    MetricName:
      type: string
      enum: [lcp, inp, cls, fcp, ttfb]
...
    DeviceClass:
      type: string
      enum: [mobile, desktop, tablet, unknown]
...
    Deployment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        environment:
          $ref: "#/components/schemas/Environment"
        buildId:
          type: string
        releaseVersion:
          type:
            - string
            - "null"
        gitRef:
          type:
            - string
            - "null"
        gitCommitSha:
          type:
            - string
            - "null"
        deployedAt:
          type: string
          format: date-time
      required: [id, siteId, environment, buildId, deployedAt]
...
    RollupPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        metricName:
          $ref: "#/components/schemas/MetricName"
        deviceClass:
          $ref: "#/components/schemas/DeviceClass"
        pageGroupKey:
          type:
            - string
            - "null"
        p75Value:
          type: number
        p50Value:
          type: number
        sampleCount:
          type: integer
        poorCount:
          type: integer
      required: [date, metricName, deviceClass, p75Value, p50Value, sampleCount, poorCount]
    SiteHealthSummary:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        siteSlug:
          type: string
        siteName:
          type: string
        failingMetrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricName"
        poorEventShare:
          type: number
      required: [siteId, siteSlug, siteName, failingMetrics, poorEventShare]
    OverviewResponse:
      type: object
      properties:
        health:
          type: array
          items:
            $ref: "#/components/schemas/SiteHealthSummary"
        trends:
          type: array
          items:
            $ref: "#/components/schemas/RollupPoint"
      required: [health, trends]
...
    DeploymentCompareMetric:
      type: object
      properties:
        metricName:
          $ref: "#/components/schemas/MetricName"
        deviceClass:
          $ref: "#/components/schemas/DeviceClass"
        pageGroupKey:
          type:
            - string
            - "null"
        baselineP75:
          type: number
        currentP75:
          type: number
        delta:
          type: number
        baselineSampleCount:
          type: integer
        currentSampleCount:
          type: integer
      required:
        - metricName
        - deviceClass
        - baselineP75
        - currentP75
        - delta
        - baselineSampleCount
        - currentSampleCount
    DeploymentCompareResponse:
      type: object
      properties:
        currentDeployment:
          $ref: "#/components/schemas/Deployment"
        baselineDeployment:
          $ref: "#/components/schemas/Deployment"
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentCompareMetric"
      required: [currentDeployment, baselineDeployment, metrics]